courses/agentic_ai/examples/w3/requirements.txt
pypdf
pdf2image
pytesseract
chromadb
tesseract
google-genai

sentence-transformers
transformers
huggingface-hub
numpy==1.26.0